requirements.txt
absl-py==1.0.0
ale-py==0.7.3
appnope==0.1.2
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
gym==0.21.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
Markdown==3.3.6
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.21
Pygments==2.10.0
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
